New_Exceptions

+++ Pl2-Project/src/project/AdminEmployee.java
@@ -39,11 +39,11 @@ public class AdminEmployee extends Employee {
                     + "\nAlter your User-Name and Password.  (Enter 7)"
                     + "\nLogOut                              (Enter 8)\n");
             System.out.printf("?: ");
-            c = input.nextInt();
-            input.nextLine();
+            c = Check.CheckNumber();
 
             if (c != 1 && c != 2 && c != 3 && c != 4 && c != 5 && c != 6 && c != 7 & c != 8) {
                 System.out.println("Invaild Input!");
+                continue;
             }
 
             switch (c) {
@@ -78,18 +78,18 @@ public class AdminEmployee extends Employee {
     private void Add_new_employee() {
         String fName;
         String lName;
-        String fullName;
+        //String fullName;
         String userName;
         String password;
         String eType = "";
         Scanner input = new Scanner(System.in);
         System.out.print("Enter employee's first name: ");
-        fName = input.nextLine();
+        fName = Check.CheckFname();
         System.out.print("Enter employee's last name: ");
-        lName = input.nextLine();
-        fullName = fName + lName;
+        lName = Check.CheckLname();
+        //fullName = fName + lName;
         System.out.print("Enter employee's username: ");
-        userName = input.nextLine();
+        userName = Check.CheckNewUsername();
         System.out.print("Enter employee's password: ");
         password = input.nextLine();
         int c;
@@ -101,11 +101,11 @@ public class AdminEmployee extends Employee {
                     + "\nSales Departement.          (Enter 4)\n"
             );
             System.out.printf("?: ");
-            c = input.nextInt();
-            input.nextLine();
+            c = Check.CheckNumber();
 
             if (c != 1 && c != 2 && c != 3 && c != 4) {
                 System.out.println("Invaild Input!");
+                continue;
             }
 
             switch (c) {
@@ -133,15 +133,19 @@ public class AdminEmployee extends Employee {
         } else if (eType.equals("I")) {
             EmployeeDB.add_employee(new InventoryEmployee(fName, lName, userName, password));
         }
-        System.out.printf("\n%sAdded!\n", fName);
+        System.out.printf("\n%s is Added!\n", fName);
     }
 
     private void Delete_an_employee() {
         int id;
         Scanner input = new Scanner(System.in);
         System.out.print("Enter the id of the employee you want to delete: ");
-        id = input.nextInt();
-        EmployeeDB.delete_employee(id);
+        id = Check.CheckID();
+        if (id == -1) {
+            System.out.println("No deletion happened");
+        } else {
+            EmployeeDB.delete_employee(id);
+        }
     }
 
     private void Update_an_employee_information() {
@@ -154,15 +158,18 @@ public class AdminEmployee extends Employee {
         String eType = "";
         Scanner input = new Scanner(System.in);
         System.out.print("Enter employee's id: ");
-        id = input.nextInt();
-        if (EmployeeDB.isExisit(id)) {
+        id = Check.CheckID();
+        if (id == -1) {
+            System.out.println("No update happened");
+        } else {
+//        if (EmployeeDB.isExist(id)) {
             System.out.print("Enter employee's first name: ");
-            fName = input.next();
+            fName = Check.CheckFname();
             System.out.print("Enter employee's last name: ");
-            lName = input.next();
+            lName = Check.CheckLname();
             //fullName=fName+lName;
             System.out.print("Enter employee's username: ");
-            userName = input.next();
+            userName = Check.CheckNewUsername();
             System.out.print("Enter employee's password: ");
             password = input.next();
             int c;
@@ -174,11 +181,11 @@ public class AdminEmployee extends Employee {
                         + "\nSales Departement.          (Enter 4)\n"
                 );
                 System.out.printf("?: ");
-                c = input.nextInt();
-                input.nextLine();
+                c = Check.CheckNumber();
 
                 if (c != 1 && c != 2 && c != 3 && c != 4) {
                     System.out.println("Invaild Input!");
+                    continue;
                 }
 
                 switch (c) {
@@ -197,9 +204,10 @@ public class AdminEmployee extends Employee {
                 }
             } while (c != 1 && c != 2 && c != 3 && c != 4);
             EmployeeDB.update_employee(id, fName, lName, userName, password, eType);
-        } else {
-            System.out.println("\nNot Found!");
         }
+//        } else {
+//            System.out.println("\nNot Found!");
+//        }
     }
 
     private void list_all_employees() {
@@ -223,20 +231,20 @@ public class AdminEmployee extends Employee {
         }
     }
 
-    void update_general_info() {
+    private void update_general_info() {
         Scanner input = new Scanner(System.in);
         System.out.printf("Enter first name : ");
-        String fname = input.next();
+        String fname = Check.CheckFname();
         System.out.printf("Enter last name : ");
-        String lname = input.next();
+        String lname = Check.CheckLname();
         EmployeeDB.update_employee_info(this.getId(), fname, lname);
     }
 
-    void update_general() {
+    private void update_general() {
 
         Scanner input = new Scanner(System.in);
         System.out.printf("Enter user name : ");
-        String username = input.next();
+        String username = Check.CheckNewUsername();
         System.out.printf("Enter password : ");
         String password = input.next();
         EmployeeDB.update_employee(this.getId(), username, password);
@@ -247,17 +255,20 @@ public class AdminEmployee extends Employee {
         ArrayList<Employee> list = new ArrayList<>();
         list = EmployeeDB.get_employees();
         System.out.printf("Enter id : ");
-        int id = input.nextInt();
+        int id = Check.CheckID();
+        if (id == -1) {
+            System.out.println("No search happened");
+        } else {
+            for (int i = 0; i < list.size(); i++) {
+                if (list.get(i).getId() == id) {
+                    System.out.println("id : \t" + list.get(i).getId());
+                    System.out.println("first name : \t" + list.get(i).getfName());
+                    System.out.println("last name : \t" + list.get(i).getlName());
+                    System.out.println("user name : \t" + list.get(i).getUserName());
+                    System.out.println("password : \t" + list.get(i).getfName());
+                    System.out.println("employee type : " + list.get(i).getEType());
 
-        for (int i = 0; i < list.size(); i++) {
-            if (list.get(i).getId() == id) {
-                System.out.println("id : \t" + list.get(i).getId());
-                System.out.println("first name : \t" + list.get(i).getfName());
-                System.out.println("last name : \t" + list.get(i).getlName());
-                System.out.println("user name : \t" + list.get(i).getUserName());
-                System.out.println("password : \t" + list.get(i).getfName());
-                System.out.println("employee type : " + list.get(i).getEType());
-
+                }
             }
         }
